[REFACTOR] SPN 수정
accConumption 부분 수정
SPN 스케줄 다음 History 에 totalAccConsumption, totalBurstTime 설정

+++ src/team/os/model/Core.java
@@ -6,10 +6,9 @@ public abstract class Core {
 	// 작업량 PCore = 2, ECore = 1
 	private int power;
 	// 누적 전력 소모량
-	private double accConsumption = 0;
+	private int accConsumption = 0;
 	// 일하고 있는지 여부
 	private boolean working = false;
-	
 	
 	public boolean isWorking() {
 		return working;
@@ -29,10 +28,10 @@ public abstract class Core {
 	public void setPower(int power) {
 		this.power = power;
 	}
-	public double getAccConsumption() {
+	public int getAccConsumption() {
 		return accConsumption;
 	}
-	public void setAccConsumption(double accConsumption) {
+	public void setAccConsumption(int accConsumption) {
 		this.accConsumption = accConsumption;
 	}
 }
